+++ playback/replayplaylist.cpp
@@ -63,7 +63,7 @@ int ReplayPlaylist::markIn(const QString& clipPath, qint64 inMs) {
     e.outMs = -1;
     e.speed = 1.0;
     m_entries.append(e);
-    return m_entries.size() - 1;
+    return static_cast<int>(m_entries.size()) - 1;
 }
 
 bool ReplayPlaylist::markOut(qint64 outMs) {
